+++ react/MiniCart.js
@@ -1,10 +1,13 @@
 import React, { Component } from 'react'
 import orderFormQuery from './graphql/orderFormQuery.gql'
 import { graphql } from 'react-apollo'
-import Button from '@vtex/styleguide/lib/Button'
+import { Button } from 'vtex.styleguide'
+import classNames from 'classnames'
+
 import CartIcon from './images/CartIcon'
 import MiniCartContent from './components/MiniCartContent'
 import { MiniCartPropTypes } from './propTypes'
+import Sidebar from './components/Sidebar'
 
 import './global.css'
 
@@ -26,6 +29,7 @@ export class MiniCart extends Component {
           widget: {
             'ui:widget': 'range',
           },
+          isLayout: true,
         },
       }
     }
@@ -37,87 +41,159 @@ export class MiniCart extends Component {
       description: 'editor.minicart.description',
       type: 'object',
       properties: {
+        type: {
+          title: 'editor.minicart.type.title',
+          type: 'string',
+          default: 'popup',
+          enum: ['popup', 'sidebar'],
+          enumNames: [
+            'editor.minicart.type.popup',
+            'editor.minicart.type.sidebar',
+          ],
+          widget: {
+            'ui:widget': 'radio',
+            'ui:options': {
+              'inline': true,
+            },
+          },
+          isLayout: true,
+        },
         showRemoveButton: {
           title: 'editor.minicart.showRemoveButton.title',
           type: 'boolean',
+          isLayout: true,
+        },
+        showDiscount: {
+          title: 'editor.minicart.showDiscount.title',
+          type: 'boolean',
+          isLayout: true,
         },
         labelMiniCartEmpty: {
           title: 'editor.minicart.labelMiniCartEmpty.title',
           type: 'string',
+          isLayout: false,
         },
         labelButtonFinishShopping: {
           title: 'editor.minicart.labelButtonFinishShopping.title',
           type: 'string',
+          isLayout: false,
         },
         enableQuantitySelector: {
           title: 'editor.minicart.enableQuantitySelector.title',
           type: 'boolean',
+          isLayout: true,
         },
         ...generatedSchema,
       },
     }
   }
 
-  constructor(props) {
-    super(props)
-    this.state = { isMouseOnButton: false, isMouseOnMiniCart: false, quantityItems: 0 }
+  state = {
+    quantityItems: 0,
+    showSideBar: false,
   }
 
   componentDidMount() {
-    document.addEventListener('item:add', () => {
-      const { quantityItems } = this.state
-      this.setState({ quantityItems: quantityItems + 1 })
+    document.addEventListener('item:add', this.handleItemAdd)
+  }
+
+  handleClickButton = () => {
+    const { hideContent, type } = this.props
+    if (!hideContent) {
+      if (type && type === 'sidebar') {
+        this.setState({
+          showSideBar: true,
+        })
+      } else {
+        location.assign('/checkout/#/cart')
+      }
+    }
+  }
+
+  handleCloseSideBarButtonClick = () => {
+    this.setState({
+      showSideBar: false,
     })
+  }
+
+  componentWillUnmount() {
+    document.removeEventListener('item:add', this.handleItemAdd)
+  }
+
+  handleItemAdd = () => {
+    this.props.data.refetch()
   }
 
   handleUpdateQuantityItems = quantity => this.setState({ quantityItems: quantity })
 
-  handleClickButton = () => location.assign('/checkout/#/cart')
-
-  handleMouseEnterButton = () => this.setState({ isMouseOnButton: true })
-
-  handleMouseLeaveButton = () => this.setState({ isMouseOnButton: false })
-
-  handleMouseEnterCartItems = () => this.setState({ isMouseOnMiniCart: true })
-
-  handleMouseLeaveCartItems = () => this.setState({ isMouseOnMiniCart: false })
-
   render() {
-    const { isMouseOnButton, isMouseOnMiniCart, quantityItems } = this.state
-    const { labelMiniCartEmpty, labelButtonFinishShopping, miniCartIconColor, showRemoveButton, enableQuantitySelector, maxQuantity, data: { orderForm } } = this.props
-    const quantity = !quantityItems && orderForm && orderForm.items ? orderForm.items.length : quantityItems
+    const { showSideBar } = this.state
+    const {
+      labelMiniCartEmpty,
+      labelButtonFinishShopping,
+      miniCartIconColor,
+      showRemoveButton,
+      showDiscount,
+      enableQuantitySelector,
+      maxQuantity,
+      data,
+      type,
+      hideContent,
+    } = this.props
+    const { orderForm } = data
+    const quantity = orderForm && orderForm.items ? orderForm.items.length : 0
+    const classes = classNames('mt3 bg-white relative', {
+      'vtex-minicart__content-container--footer-small': !showDiscount,
+      'vtex-minicart__content-container--footer-large': showDiscount,
+    })
+
     return (
-      <div className="relative fr" >
+      <div className="vtex-minicart relative fr">
         <Button
+          variation="tertiary"
           icon
           onClick={this.handleClickButton}
-          onMouseEnter={this.handleMouseEnterButton}
-          onMouseLeave={this.handleMouseLeaveButton}>
+        >
           <CartIcon fillColor={miniCartIconColor} />
           {quantity > 0 && <span className="vtex-minicart__bagde mt1 mr1">
             {quantity}
           </span>}
         </Button>
-        {
-          (isMouseOnMiniCart || isMouseOnButton) &&
-          <div
-            className="vtex-minicart__box absolute right-0 z-max flex flex-colunm"
-            onMouseLeave={this.handleMouseLeaveCartItems}
-            onMouseEnter={this.handleMouseEnterCartItems}>
-            <div className="vtex-minicart__arrow-up absolute top-0 right-0 shadow-3">
+        {!hideContent && ((type && type === 'sidebar') ? (showSideBar &&
+          <Sidebar
+            onBackClick={this.handleCloseSideBarButtonClick}
+          >
+            <MiniCartContent
+              large
+              data={data}
+              onUpdateItemsQuantity={this.handleUpdateQuantityItems}
+              showRemoveButton={showRemoveButton}
+              showDiscount={showDiscount}
+              labelMiniCartEmpty={labelMiniCartEmpty}
+              labelButton={labelButtonFinishShopping}
+              enableQuantitySelector={enableQuantitySelector}
+              maxQuantity={maxQuantity}
+            />
+          </Sidebar>)
+          : (!showSideBar &&
+            <div className="vtex-minicart__box absolute right-0 z-max flex flex-colunm">
+              <div className="shadow-3">
+                <div className="vtex-minicart__arrow-up absolute top-0 right-0 shadow-3" />
+                <div className={classes}>
+                  <MiniCartContent
+                    data={data}
+                    onUpdateItemsQuantity={this.handleUpdateQuantityItems}
+                    showRemoveButton={showRemoveButton}
+                    showDiscount={showDiscount}
+                    labelMiniCartEmpty={labelMiniCartEmpty}
+                    labelButton={labelButtonFinishShopping}
+                    enableQuantitySelector={enableQuantitySelector}
+                    maxQuantity={maxQuantity}
+                  />
+                </div>
+              </div>
             </div>
-            <div className="shadow-3 mt3">
-              <MiniCartContent
-                data={this.props.data}
-                onUpdateItemsQuantity={this.handleUpdateQuantityItems}
-                showRemoveButton={showRemoveButton}
-                labelMiniCartEmpty={labelMiniCartEmpty}
-                labelButton={labelButtonFinishShopping}
-                enableQuantitySelector={enableQuantitySelector}
-                maxQuantity={maxQuantity} />
-            </div>
-          </div>
-        }
+          ))}
       </div>
     )
   }