+++ react/components/MiniCartContent.js
@@ -1,22 +1,26 @@
-import React, { Component } from 'react'
+import React, { Component, Fragment } from 'react'
 import PropTypes from 'prop-types'
 import { graphql } from 'react-apollo'
 import { injectIntl, intlShape } from 'react-intl'
+import { reduceBy, values } from 'ramda'
+import classNames from 'classnames'
+
+import { Button, Spinner } from 'vtex.styleguide'
+import ProductPrice from 'vtex.store-components/ProductPrice'
+
 import updateItemsMutation from '../graphql/updateItemsMutation.gql'
 import orderFormQuery from '../graphql/orderFormQuery.gql'
-import MiniCartItem from './MiniCartItem'
-import Button from '@vtex/styleguide/lib/Button'
-import ProductPrice from 'vtex.store-components/ProductPrice'
-import Spinner from '@vtex/styleguide/lib/Spinner'
-import { MiniCartPropTypes } from '../propTypes'
 
-import '../global.css'
+import MiniCartItem from './MiniCartItem'
+import { MiniCartPropTypes } from '../propTypes'
 
 /**
  * Minicart content component
  */
 class MiniCartContent extends Component {
   static propTypes = {
+    /* Set the mini cart content size */
+    large: PropTypes.bool,
     /* Function to be called when an item is removed */
     onUpdateItemsQuantity: PropTypes.func,
     /* Mutate function */
@@ -30,11 +34,25 @@ class MiniCartContent extends Component {
     showRemoveButton: MiniCartPropTypes.showRemoveButton,
     enableQuantitySelector: MiniCartPropTypes.enableQuantitySelector,
     maxQuantity: MiniCartPropTypes.maxQuantity,
+    showDiscount: MiniCartPropTypes.showDiscount,
   }
 
   constructor(props) {
     super(props)
     this.state = { showSpinner: false }
+  }
+
+  sumItemsPrice = items => {
+    let sum = 0
+    items.forEach(item => {
+      sum += (item.listPrice * item.quantity)
+    })
+    return sum
+  }
+
+  calculateDiscount = (items, totalPrice) => {
+    const liquidPrice = this.sumItemsPrice(items)
+    return totalPrice - liquidPrice
   }
 
   handleClickButton = () => location.assign('/checkout/#/cart')
@@ -89,43 +107,90 @@ class MiniCartContent extends Component {
   }
 
   renderWithoutItems = label => (
-    <div className="vtex-minicart__item pa4 flex items-center justify-center relative bg-white">
+    <div className="vtex-minicart__item pa4 flex items-center justify-center relative bg-white pt9">
       <span className="f5">{label}</span>
     </div>
   )
 
-  renderMiniCartWithItems = (orderForm, label, showRemoveButton, enableQuantitySelector, maxQuantity, showSpinner) => (
-    <div className="flex flex-column relative" >
-      <div className="bg-white">
-        <div className="vtex-minicart__content pr4 pl4 overflow-auto overflow-x-hidden">
-          {orderForm.items.map(item => (
-            <div className="flex flex-row" key={item.id}>
-              <MiniCartItem
-                {...item}
-                removeItem={this.onRemoveItem}
-                updateItem={this.onUpdateItems}
-                showRemoveButton={showRemoveButton}
-                enableQuantitySelector={enableQuantitySelector}
-                maxQuantity={maxQuantity} />
-            </div>
+  renderMiniCartWithItems = (
+    orderForm,
+    label,
+    labelDiscount,
+    showRemoveButton,
+    showDiscount,
+    enableQuantitySelector,
+    maxQuantity,
+    showSpinner,
+    large
+  ) => {
+    const items = values(
+      reduceBy(
+        (acc, item) =>
+          acc ? { ...acc, quantity: acc.quantity + item.quantity } : item,
+        undefined,
+        item => item.id,
+        orderForm.items
+      )
+    )
+
+    const classes = classNames('vtex-minicart__content overflow-x-hidden h-100', {
+      'vtex-minicart__content--small': !large,
+      'vtex-minicart__content--large': large,
+      'vtex-minicart__content-large--footer-small': large && !showDiscount,
+      'vtex-minicart__content-large--footer-large': large && showDiscount,
+      'overflow-y-scroll': items.length > 3 && !large,
+      'overflow-y-hidden': items.length <= 3 && !large,
+    })
+
+    const discount = this.calculateDiscount(orderForm.items, orderForm.value)
+
+    return (
+      <Fragment>
+        <div className={classes}>
+          {items.map(item => (
+            <MiniCartItem
+              {...item}
+              key={item.id}
+              large
+              removeItem={this.onRemoveItem}
+              updateItem={this.onUpdateItems}
+              showRemoveButton={showRemoveButton}
+              enableQuantitySelector={enableQuantitySelector}
+              maxQuantity={maxQuantity}
+            />
           ))}
         </div>
-        <div className="fl pa4">
-          <Button primary onClick={this.handleClickButton}>{label}</Button>
-        </div>
-        <div className="flex flex-row fr pt4 mt2 mr4">
-          {showSpinner &&
-            <Spinner size={18} />
+        <div className="absolute bottom-0 w-100 bg-white flex flex-column pa4 bt b--silver pt4">
+          {showDiscount && <div className="vtex-minicart__content-discount w-100 mb4">
+            <span className="ttu b">{labelDiscount}</span>
+            <div className="fr">
+              <ProductPrice
+                sellingPrice={discount}
+                listPrice={discount}
+                showLabels={false}
+                showListPrice={false}
+              />
+            </div>
+          </div>
           }
-          <ProductPrice
-            sellingPrice={orderForm.value}
-            listPrice={orderForm.value}
-            showLabels={false}
-            showListPrice={false} />
+          <div className="relative">
+            <div className="fl">
+              <Button variation="primary" size="small" onClick={this.handleClickButton}>{label}</Button>
+            </div>
+            <div className="absolute right-0 mt3 flex flex-row">
+              {showSpinner && <Spinner size={18} />}
+              <ProductPrice
+                sellingPrice={orderForm.value}
+                listPrice={orderForm.value}
+                showLabels={false}
+                showListPrice={false}
+              />
+            </div>
+          </div>
         </div>
-      </div>
-    </div >
-  )
+      </Fragment>
+    )
+  }
 
   renderLoading = () => (
     <div className="vtex-minicart__item pa4 flex items-center justify-center relative bg-white">
@@ -133,26 +198,43 @@ class MiniCartContent extends Component {
     </div>
   )
 
-  componentDidMount() {
-    this.props.data.refetch().then(() => {
-      this.props.onUpdateItemsQuantity(this.props.data.orderForm.items.length)
-    })
-  }
-
   render() {
-    const { data, labelMiniCartEmpty, labelButton, intl, showRemoveButton, enableQuantitySelector, maxQuantity } = this.props
+    const {
+      data,
+      labelMiniCartEmpty,
+      labelButton,
+      intl,
+      showRemoveButton,
+      showDiscount,
+      enableQuantitySelector,
+      maxQuantity,
+      large,
+    } = this.props
     const { showSpinner } = this.state
-    let content
-    if (data.loading) {
-      content = this.renderLoading()
-    } else if (!data.orderForm || !data.orderForm.items.length) {
-      const label = labelMiniCartEmpty || intl.formatMessage({ id: 'minicart-empty' })
-      content = this.renderWithoutItems(label)
-    } else {
-      const label = labelButton || intl.formatMessage({ id: 'finish-shopping-button-label' })
-      content = this.renderMiniCartWithItems(data.orderForm, label, showRemoveButton, enableQuantitySelector, maxQuantity, showSpinner)
+
+    if (!data || data.loading) {
+      return this.renderLoading()
     }
-    return content
+
+    if (!data.orderForm || !data.orderForm.items.length) {
+      const label = labelMiniCartEmpty || intl.formatMessage({ id: 'minicart-empty' })
+      return this.renderWithoutItems(label)
+    }
+
+    const label = labelButton || intl.formatMessage({ id: 'finish-shopping-button-label' })
+    const labelDiscount = intl.formatMessage({ id: 'minicart-content-footer-discount' })
+
+    return this.renderMiniCartWithItems(
+      data.orderForm,
+      label,
+      labelDiscount,
+      showRemoveButton,
+      showDiscount,
+      enableQuantitySelector,
+      maxQuantity,
+      showSpinner,
+      large
+    )
   }
 }