+++ react/components/MiniCartItem.js
@@ -1,15 +1,11 @@
 import React, { Component } from 'react'
 import PropTypes from 'prop-types'
 import { Link } from 'render'
-import CloseIcon from '@vtex/styleguide/lib/icon/Close'
-import Button from '@vtex/styleguide/lib/Button'
-import Spinner from '@vtex/styleguide/lib/Spinner'
+import { Button, Spinner, IconClose } from 'vtex.styleguide'
 import ProductName from 'vtex.store-components/ProductName'
 import ProductPrice from 'vtex.store-components/ProductPrice'
 import QuantitySelector from 'vtex.store-components/QuantitySelector'
 import { MiniCartPropTypes } from '../propTypes'
-
-import '../global.css'
 
 /**
  * Minicart item component
@@ -18,6 +14,8 @@ export default class MiniCartItem extends Component {
   static propTypes = {
     /* Item's id */
     id: PropTypes.string.isRequired,
+    /* Item's type */
+    type: PropTypes.string,
     /* Item's image */
     imageUrl: PropTypes.string.isRequired,
     /* Item's name */
@@ -88,55 +86,54 @@ export default class MiniCartItem extends Component {
     const { isRemovingItem } = this.state
 
     return (
-      <div className="relative">
+      <div className="vtex-minicart__item relative pv3 w-100 ph4 mb2">
         <Link
           className="pointer link black-90"
           page={'store/product'}
           params={{ slug: this.getItemId(detailUrl) }}>
-          <div className="vtex-minicart__item flex flex-row relative bb b--silver mt4">
-            <div className="vtex-minicart__img-container">
-              <img className="vtex-minicart__item-image" src={imageUrl} alt={name} />
+          <div className="relative bb b--silver">
+            <div className="w-100 mb2">
+              <ProductName name={name} />
+              <div className="f7">
+                <ProductName name={skuName} />
+              </div>
             </div>
-            <div className="ml3 relative">
-              <div>
-                <div className="vtex-minicart__item-name mt3 tl overflow-hidden">
-                  <ProductName name={name} />
-                </div>
-                <div className="vtex-minicart__sku-name tl absolute">
-                  <div className="f7 dark-gray">
-                    <ProductName name={skuName} />
-                  </div>
-                </div>
+            <div className="vtex-minicart__item-footer relative flex flex-row pb2 items-center w-100">
+              <div className="vtex-minicart__img-container">
+                <img className="vtex-minicart__item-image" src={imageUrl} alt={name} />
+              </div>
+              <div className="justify-end absolute right-0 pt6">
+                <ProductPrice
+                  sellingPrice={sellingPrice * quantity}
+                  listPrice={listPrice * quantity}
+                  showLabels={false}
+                  showListPrice={false} />
               </div>
             </div>
           </div>
         </Link>
-        <div className="absolute right-0 bottom-0 mb4 flex flex-column">
-          <ProductPrice
-            sellingPrice={sellingPrice * quantity}
-            listPrice={listPrice * quantity}
-            showLabels={false}
-            showListPrice={false} />
-        </div>
         {enableQuantitySelector &&
-          <div className="absolute bottom-0 right-0 mb7 pb2">
-            <QuantitySelector maxQuantity={maxQuantity} currentQuantity={quantity} onQuantityChange={this.handleQuantityChange} />
+          <div className="absolute bottom-0 left-0 mb4 pl9">
+            <QuantitySelector
+              maxQuantity={maxQuantity}
+              currentQuantity={quantity}
+              onQuantityChange={this.handleQuantityChange}
+            />
           </div>
         }
-        {
-          (showRemoveButton && !isRemovingItem) &&
-          <div className="vtex-minicart-item__remove-btn absolute right-0 top-0 mt4">
-            <Button onClick={(e) => this.onClickRemove(id, e)}>
-              <CloseIcon size={12} color="#BDBDBD" />
+        {(showRemoveButton && !isRemovingItem) && (
+          <div className="vtex-minicart-item__remove-btn absolute right-0 top-0 mr4">
+            <Button icon variation="tertiary" onClick={(e) => this.onClickRemove(id, e)}>
+              <IconClose size={20} color="#BDBDBD" />
             </Button>
           </div>
-        }
-        {
-          (showRemoveButton && isRemovingItem) &&
-          <div className="vtex-minicart-item__remove-btn absolute right-0 top-0 flex items-center justify-center mt5 pt2">
+        )}
+        {(showRemoveButton && isRemovingItem) && (
+          <div
+            className="vtex-minicart-item__remove-btn absolute right-0 top-0 flex items-center justify-center mt3 mr4">
             <Spinner size={20} />
           </div>
-        }
+        )}
       </div>
     )
   }