+++ react/propTypes.js
@@ -1,6 +1,10 @@
 import PropTypes from 'prop-types'
 
 export const MiniCartPropTypes = {
+  /* Set the minicart type */
+  type: PropTypes.string,
+  /* Set the content visibility */
+  hideContent: PropTypes.bool,
   /* Label that wiil appear when the minicart is empty */
   labelMiniCartEmpty: PropTypes.string,
   /* Finish shopping button label */
@@ -9,6 +13,8 @@ export const MiniCartPropTypes = {
   miniCartIconColor: PropTypes.string,
   /* Set the visibility of remove button */
   showRemoveButton: PropTypes.bool,
+  /* Set the discount visibility */
+  showDiscount: PropTypes.bool,
   /* Set the visibility of the Quantity Selector */
   enableQuantitySelector: PropTypes.bool,
   /* Max quantity for the Quantity Selector */