+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Minicart SideBar type.
+
+### Fixed
+- Fix duplicate items showing when changing the quantity.
 
 ## [0.6.0] - 2018-6-20
 ### Added